Moved files around :) and formatting :)
