+++ Assets/Menus/MainMenu/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0f87c0eca156e41d496ec9da54c7b9bd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
