+++ Assets/Time/Scripts/TimeUI.cs
@@ -1,0 +1,35 @@
+using TMPro;
+using UnityEngine;
+
+namespace Time.Scripts
+{
+    public class TimeUI : MonoBehaviour
+    {
+        public TextMeshProUGUI timeText;
+        public TextMeshProUGUI dayText;
+
+        private void OnEnable()
+        {
+            WorldTime.OnMinuteChanged += UpdateTime;
+            WorldTime.OnHourChanged += UpdateTime;
+            WorldTime.OnDayChanged += UpdateDay;
+        }
+
+        private void OnDisable()
+        {
+            WorldTime.OnMinuteChanged -= UpdateTime;
+            WorldTime.OnHourChanged -= UpdateTime;
+            WorldTime.OnDayChanged -= UpdateDay;
+        }
+
+        private void UpdateDay()
+        {
+            dayText.text = "Day " + WorldTime.Day;
+        }
+
+        private void UpdateTime()
+        {
+            timeText.text = WorldTime.Hour.ToString("00") + ":" + WorldTime.Minute.ToString("00");
+        }
+    }
+}
