+++ Assets/World/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ab382bf67a4634969b7030af86659102
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
