+++ UserSettings/EditorUserSettings.asset
@@ -6,10 +6,10 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedSceneGuid-0:
-      value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
+      value: 5a505657000d5b025a59082042270a44434f1b7b7f7f76692b7a4b60e3e63768
       flags: 0
     RecentlyUsedSceneGuid-1:
-      value: 5a505657000d5b025a59082042270a44434f1b7b7f7f76692b7a4b60e3e63768
+      value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
